+++ CHANGELOG.md
@@ -9,9 +9,24 @@ This project adheres to [Semantic Versioning], with the exception that minor rel
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-02-05
+
+### Added
+
+- added ability to schedule jumps ([#319]) ([**@aaronleesander**])
+- added Bose-Hubbard Hamiltonian option ([#309]) ([**@lucello**])
+- added multi-threading setting for noise-free ones ([#316]) ([**@aaronleesander**])
+- Minor improvements to TDVP performance ([#311]) ([**@aaronleesander**])
+- ⚡️ Improve Lanczos iteration and use numba for significant speedup ([#310]) ([**@aaronleesander**])
+
 ### Changed
 
+- ♻️ Change Pauli summation to use finite state machine construction ([#308]) ([**@aaronleesander**])
 - 🔧 Replace `mypy` with `ty` ([#304]) ([**@denialhaag**])
+
+### Removed
+
+- 👷‍♂️ Stop testing on x86 macOS systems ([#310]) ([**@aaronleesander**])
 
 ## [0.3.3] - 2026-01-12
 
@@ -53,13 +68,20 @@ _📚 Refer to the [GitHub Release Notes](https://github.com/munich-quantum-tool
 
 <!-- Version links -->
 
-[Unreleased]: https://github.com/munich-quantum-toolkit/yaqs/compare/v0.3.3...HEAD
+[Unreleased]: https://github.com/munich-quantum-toolkit/yaqs/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/munich-quantum-toolkit/yaqs/releases/tag/v0.4.0
 [0.3.3]: https://github.com/munich-quantum-toolkit/yaqs/releases/tag/v0.3.3
 [0.3.2]: https://github.com/munich-quantum-toolkit/yaqs/releases/tag/v0.3.2
 [0.3.1]: https://github.com/munich-quantum-toolkit/yaqs/releases/tag/v0.3.1
 
 <!-- PR links -->
 
+[#319]: https://github.com/munich-quantum-toolkit/yaqs/pull/319
+[#316]: https://github.com/munich-quantum-toolkit/yaqs/pull/316
+[#311]: https://github.com/munich-quantum-toolkit/yaqs/pull/311
+[#310]: https://github.com/munich-quantum-toolkit/yaqs/pull/310
+[#309]: https://github.com/munich-quantum-toolkit/yaqs/pull/309
+[#308]: https://github.com/munich-quantum-toolkit/yaqs/pull/308
 [#304]: https://github.com/munich-quantum-toolkit/yaqs/pull/304
 [#298]: https://github.com/munich-quantum-toolkit/yaqs/pull/298
 [#216]: https://github.com/munich-quantum-toolkit/yaqs/pull/216
@@ -78,6 +100,7 @@ _📚 Refer to the [GitHub Release Notes](https://github.com/munich-quantum-tool
 [**@denialhaag**]: https://github.com/denialhaag
 [**@aaronleesander**]: https://github.com/aaronleesander
 [**@thilomueller**]: https://github.com/thilomueller
+[**@lucello**]: https://github.com/lucello
 
 <!-- General links -->